TRIPLEX-820 Remove Island stories file as part of restructuring efforts to streamline component documentation and testing. This change eliminates outdated examples and focuses on maintaining relevant stories for visual testing.
